warnings removing

+++ stly.h
@@ -256,15 +256,15 @@ size_t Matrix<T>::get_cols() const
 template<typename T>
 void Matrix<T>::rotate()
 {
-	for(int i=0; i<rows; i++)
+	for(size_t i=0; i<rows; i++)
 	{
-		for(int j=i+1; j<cols; j++)
+		for(size_t j=i+1; j<cols; j++)
 			std::swap(arr[i*cols + j], arr[j*cols + i]);
 	}
 
-	for(int i=0; i<rows; i++)
+	for(size_t i=0; i<rows; i++)
 	{
-		for(int j=0; j<cols/2; j++)
+		for(size_t j=0; j<cols/2; j++)
 		{
 			std::swap(arr[i*cols + j], arr[i*cols + cols-j-1]);
 		}
